troise-ai/pyproject.toml
[project]
name = "troise-ai"
version = "0.1.0"
description = "TROISE AI - Personal AI Augmentation System"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",

    # Strands Agents (agent loop, tools, multi-agent)
    "strands-agents>=0.1.0",

    # Strands model backends
    "ollama>=0.4.0",
    "openai>=1.50.0",

    # DynamoDB
    "aiobotocore>=2.15.0",
    "aioboto3>=13.0.0",

    # File operations
    "aiofiles>=24.0.0",

    # SSH for DGX Spark
    "asyncssh>=2.18.0",

    # Utilities
    "pyyaml>=6.0",
    "aiohttp>=3.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",

    # RAG: Web content extraction & chunking
    "beautifulsoup4>=4.12.0",
    "langchain-text-splitters>=0.3.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
